requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
allure_python_commons
clickhouse_sqlalchemy
Flask
jsonpath
pandas
paramiko
pymongo
PyMySQL
PyQt5
pytest
PyYAML
redis
requests
SQLAlchemy
urllib3
faker
